+++ libs/dot-components/src/lib/components/skeleton/Skeleton.tsx
@@ -1,14 +1,11 @@
 import React from 'react';
-import { Skeleton } from '@material-ui/lab';
 import { CommonProps } from '../CommonProps';
 import { useStylesWithRootClass } from '../useStylesWithRootClass';
+import { rootClassName, StyledSkeleton } from './Skeleton.styles';
 
-export type SkeletonAnimationType = 'pulse' | 'wave' | false;
 export type SkeletonVariantType = 'circle' | 'rect' | 'text';
 
 export interface SkeletonProps extends CommonProps {
-  /** type of animation that will displayed on the skeleton */
-  animation?: SkeletonAnimationType;
   /** component(s) inside a skeleton will constrain the size/shape */
   children?: JSX.Element[] | JSX.Element;
   /** Determine height of skeleton */
@@ -19,11 +16,7 @@ export interface SkeletonProps extends CommonProps {
   variant?: SkeletonVariantType;
 }
 
-/**
- * @experimental This component is still in development
- */
 export const DotSkeleton = ({
-  animation = 'wave',
   children,
   className,
   'data-testid': dataTestId,
@@ -31,10 +24,10 @@ export const DotSkeleton = ({
   width,
   variant,
 }: SkeletonProps) => {
-  const rootClasses = useStylesWithRootClass('dot-skeleton', className);
+  const rootClasses = useStylesWithRootClass(rootClassName, className);
   return (
-    <Skeleton
-      animation={animation}
+    <StyledSkeleton
+      animation="wave"
       classes={{ root: rootClasses }}
       data-testid={dataTestId}
       height={height}
@@ -42,8 +35,6 @@ export const DotSkeleton = ({
       variant={variant}
     >
       {children}
-    </Skeleton>
+    </StyledSkeleton>
   );
 };
-
-export default DotSkeleton;